Generalize style and environment in DOMRenderer
